requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.2
pydantic-settings==2.6.1

# Home Assistant integration
homeassistant==2024.1.0
aiohttp==3.11.2
websockets==13.1
python-dateutil==2.9.0.post0

# AI Providers
openai==1.54.0
anthropic==0.49.0
google-generativeai==0.8.3
groq==0.11.0
cerebras-cloud-sdk==0.0.3
deepseek==1.0.0

# Telegram Integration
python-telegram-bot==21.3
python-dotenv==1.0.1

# Database & Memory
chromadb==0.5.11
sentence-transformers==3.3.1
numpy==1.26.4
pandas==2.2.3

# Voice & Audio
speechrecognition==3.10.4
pyaudio==0.2.14
pydub==0.25.1
gtts==2.5.3

# Analytics & ML
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Utilities
pyyaml==6.0.2
aiofiles==24.1.0
watchdog==5.0.3
tenacity==9.0.0
rich==13.9.4
click==8.1.7
jinja2==3.1.4

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Frontend (for serving)
httpx==0.27.2
